# swagger.yml
basePath: /api/v1
consumes:
- application/json
definitions:
  BaseResponse:
    properties:
      error:
        description: An error message describing what went wrong
        type: string
      ok:
        description: Whether the request was successful or not
        type: boolean
    required:
    - error
    - ok
    type: object
  FieldErrorResponse:
    allOf:
    - properties:
        error:
          description: An error message describing what went wrong
          type: string
        ok:
          description: Whether the request was successful or not
          type: boolean
      required:
      - error
      - ok
      type: object
    - properties:
        field_errors:
          description: A mapping of field names and errors
          type: object
      required:
      - field_errors
      type: object
  ListUserResponse:
    allOf:
    - properties:
        error:
          description: An error message describing what went wrong
          type: string
        ok:
          description: Whether the request was successful or not
          type: boolean
      required:
      - error
      - ok
      type: object
    - properties:
        count:
          description: The number of users that were returned
          format: int64
          type: integer
        users:
          description: A list of user objects
          items:
            $ref: '#/definitions/User'
          type: array
      required:
      - count
      - users
      type: object
  User:
    properties:
      created_at:
        format: date-time
        type: string
      email:
        type: string
      id:
        format: int64
        type: integer
      updated_at:
        format: date-time
        type: string
      username:
        type: string
    type: object
  UserResponse:
    allOf:
    - properties:
        error:
          description: An error message describing what went wrong
          type: string
        ok:
          description: Whether the request was successful or not
          type: boolean
      required:
      - error
      - ok
      type: object
    - properties:
        user:
          $ref: '#/definitions/User'
      required:
      - user
      type: object
host: localhost
info:
  description: Lorem ipsum
  title: Test API
  version: 0.0.1
paths:
  /users:
    get:
      operationId: listUsers
      responses:
        "200":
          $ref: '#/responses/listUserResponse'
      tags:
      - users
    post:
      operationId: createUser
      responses:
        "200":
          $ref: '#/responses/userResponse'
      tags:
      - users
  /users/{id}:
    get:
      consumes:
      - application/json
      operationId: getUser
      parameters:
      - in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/userResponse'
        "404":
          $ref: '#/responses/baseResponse'
        "500":
          $ref: '#/responses/baseResponse'
      tags:
      - users
produces:
- application/json
responses:
  baseResponse:
    description: ""
    schema:
      $ref: '#/definitions/BaseResponse'
  fieldErrorResponse:
    description: ""
    schema:
      $ref: '#/definitions/FieldErrorResponse'
  listUserResponse:
    description: ""
    schema:
      $ref: '#/definitions/ListUserResponse'
  userResponse:
    description: ""
    schema:
      $ref: '#/definitions/UserResponse'
schemes:
- http
- https
security:
- api_key: []
securityDefinitions:
  api_key:
    in: header
    name: KEY
    type: apiKey
swagger: "2.0"